Fix bad messages sent from status system to clients.
Messages sent from the status system to clients were not being
encoded with websocket framing and this was causing them to
immediately disconnect when receiving.

Pack messages from status system in Hybi before sending.

Also fixed connection ids not being updated properly because they
were not set before the ident call and it was using an undefined
value, causing all connections by id to not be found.