[9/13] Remove excess ConnectionPtr instances
In order to help debugging some reference counting bugs all of
the extra reference counted instances of connections was removed.

In retrospect none of these were ever needed as reference counting
the connections is only helpful if you are storing the connection
for longer than the scope of the function, or you are passing it
between threads.

Some connection close related changes managed to sneak into this
commit, but they are related to the next commit. They do not impact
if this commit builds and runs.