[11/13] Move close to send thread
In order to avoid race conditions with close->write actions
move all close functionality into the send thread.

There are a few minor changes to send logic in order to fix some
outstanding bugs with the send implementation, such as stalling on
a single connection if it has a very large queue that keeps being
filled.

This fixes some minor ordering related to write notification by
moving the check closer to the enqueue action, which prevents some
failures to queue when it should queue, and some double queues when
only a single queue event is needed.