[12/13] Prevent race on smart pointer inner pointer
The inner pointer of the intrusive pointer was racing between
threads because the smart pointer was being irresponsibly passed
between threads. The queue protects the new created copy and it
protects the underlying object. This is listed as a race in TSAN
but I can't find any evidence that it actually races within the
implementation details of the queue itself.